ADD : coeur fav persistant + message NoFav

+++ src/components/Card.js
@@ -1,22 +1,31 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 const Card = ({ recipe, isFavorite, onRemove }) => {
+    const [isFav, setIsFav] = useState(isFavorite);
     const limitedDishTypes = recipe.dishTypes.slice(0, 3);
     const scoreOutOfTen = (parseInt(recipe.spoonacularScore) / 10).toFixed(1);
     const defaultImage = '/assets/images/food-api.jpg';
     const navigate = useNavigate();
 
+    useEffect(() => {
+        const favorites = JSON.parse(localStorage.getItem('favRecipe')) || [];
+        setIsFav(favorites.some(fav => fav.id === recipe.id));
+    }, [recipe.id]);
+
     const handleFavClick = () => {
-        if (isFavorite && onRemove) {
-            onRemove(recipe);
+        const favorites = JSON.parse(localStorage.getItem('favRecipe')) || [];
+        if (isFav) {
+            if (onRemove) onRemove(recipe);
+            const updatedFavorites = favorites.filter(fav => fav.id !== recipe.id);
+            localStorage.setItem('favRecipe', JSON.stringify(updatedFavorites));
         } else {
-            const favorites = JSON.parse(localStorage.getItem('favRecipe')) || [];
             if (!favorites.some(fav => fav.id === recipe.id)) {
                 favorites.push(recipe);
+                localStorage.setItem('favRecipe', JSON.stringify(favorites));
             }
-            localStorage.setItem('favRecipe', JSON.stringify(favorites));
         }
+        setIsFav(!isFav);
     };
 
     const handleSeeRecipe = () => {
@@ -35,9 +44,9 @@ const Card = ({ recipe, isFavorite, onRemove }) => {
                             height="32"
                             viewBox="0 0 24 24"
                             onClick={handleFavClick}
-                            fill="currentColor"
+                            className={`${isFav ? 'filled animate' : ''}`}
                         >
-                            {isFavorite ? (
+                            {isFav ? (
                                 <path d="M1 4.27L2.28 3L20 20.72L18.73 22l-3.55-3.56l-1.73 1.59L12 21.35l-1.45-1.32C5.4 15.36 2 12.27 2 8.5c0-.95.23-1.83.63-2.6zM7.5 3c1.74 0 3.41.81 4.5 2.08C13.09 3.81 14.76 3 16.5 3C19.58 3 22 5.41 22 8.5c0 2.57-1.58 4.82-4.21 7.47L5.27 3.45C5.95 3.16 6.7 3 7.5 3" />
                             ) : (
                                 <path d="m12.1 18.55l-.1.1l-.11-.1C7.14 14.24 4 11.39 4 8.5C4 6.5 5.5 5 7.5 5c1.54 0 3.04 1 3.57 2.36h1.86C13.46 6 14.96 5 16.5 5c2 0 3.5 1.5 3.5 3.5c0 2.89-3.14 5.74-7.9 10.05M16.5 3c-1.74 0-3.41.81-4.5 2.08C10.91 3.81 9.24 3 7.5 3C4.42 3 2 5.41 2 8.5c0 3.77 3.4 6.86 8.55 11.53L12 21.35l1.45-1.32C18.6 15.36 22 12.27 22 8.5C22 5.41 19.58 3 16.5 3" />
